Remove exposition
> This is effectively the simplest possible design that will work.

Not that important to say.

> Previous design ideas for using cryptographically generated session ids, either using hash chains or public key encryption, were dismissed due to their inefficient designs. 

Same.

> Note that a client always has the chance to fall back to a full handshake or more precisely to a handshake that uses session resumption.

Why this is a relevant statement is baffling.  The point that is being made is that more complex designs might allow a connection to continue in the face of path changes (and NAT rebindings), but these are far too complex and in cases where a connection ID is insufficient, a new connection is probably easier to manage.

But you don't need to say all of that.